requirements.txt
python-binance==1.0.19
pandas>=2.2.0
numpy>=1.26.0
ta-lib>=0.4.28
python-dotenv==1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
flask>=3.0.0